requirements.txt
selenium==4.18.1
Pillow==10.2.0
argparse==1.4.0